+++ packages/subfooter/src/Blocks/ButtonBlock.js
@@ -15,7 +15,7 @@ const ButtonBlock = ({
   return (
     <div className="u-margin-bottom">
       <h5 className="u-margin-top-xs u-margin-bottom">{title}</h5>
-      <Button type="primary" title={buttonTitle} onClick={() => redirect(url)} iconLeft={icon}>{children}</Button>
+      <Button type="primary" role="link" title={buttonTitle} onClick={() => redirect(url)} iconLeft={icon}>{children}</Button>
     </div>
   );
 };